Re-factor record factory to make it more reusable.
